inference/baseline/requirements.txt
sglang
pyyaml
# torch is intentionally omitted to support custom accelerator builds (MUSA, ROCm, CANN)
